Allow for multiple hash 128 output types
This commit repurposes the Hash128 method with the Hash128x64 implementation and returns an output type that can convert the hash to a pair of uint64, a byte array or write the hash to a predefined byte array.  Nothing in our code base is using Hash128  and we're not currently using  any big-endian architectures so this can safely be removed.

This effectively combines Hash128x64 and WriteHash128x64.  Those functions now delegate to Hash128 and are left in for backwards compatibility