Improve 128-bit hash performance
This commit makes some improvements to the 128-bit hash performance, mainly to the Hash128Writer but also to the standalone Hash128 function

Prior to this, the writer would copy all incoming data in to a 16-byte buffer and process it block by block, which incurred a lot of memmmove overhead.  This has been changed to unsafe read the incoming data in 16-byte blocks and only copying the remainder to a tail buffer.  This ends up being much more efficient for larger blocks of input as well as inputs that are sized in to 16-byte blocks.

The other change here is to use `bits.RotateLeft64` where appropriate instead of hand-rolling the code for this.  Using `bits.RotateLeft64` leverages assembly intrinsics (ROLQ) where available which speeds this up

The changes to HashWriter128 are backwards compatible but new methods have been added (AddString/AddBytes) that allow callers not to have to bother with an error result.  The Sum128() method was added and returns a Hash128Value, similar to the standalone hash function

Some tests & benchmarks from the twmb/murmur3 project were borrowed to test edge cases around the handling of tail bytes in the writer.